+++ Source/Engine/Systems/Renderer/Renderer.h
@@ -20,7 +20,7 @@ namespace Engine
 
 		virtual std::unique_ptr<CubeMapController>& GetCubeMapController() = 0;
 
-		virtual void UploadMeshToMegaBuffer(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices, MeshBufferData& meshData) = 0;
+		virtual void UploadMeshToMegaBuffer(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices, MeshBufferData& meshData) = 0;
 
 		// For consistent UI scaling across the whole engine:
