+++ Source/Engine/Systems/Renderer/Vulkan/VulkanRenderer.cpp
@@ -299,7 +299,7 @@ namespace Engine
 		return 0;
 	}
 
-	void VulkanRenderer::UploadMeshToMegaBuffer(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices, MeshBufferData& meshData)
+	void VulkanRenderer::UploadMeshToMegaBuffer(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices, MeshBufferData& meshData)
 	{
 		if (indexDraw) indexDraw->UploadMeshToMegaBuffer(vertices, indices, meshData);
 	}
